refactor: Rename `compose_builders` package to `compose`
