+++ internal/builders/compose/mongodb_builder.go
@@ -1,4 +1,4 @@
-package composebuilders
+package compose
 
 import (
 	"fmt"
@@ -8,9 +8,9 @@ import (
 	"github.com/manifoldco/promptui"
 )
 
-func MysqlBuilderComposefile() error {
-	compose := setMysqlSettings()
-	path := fmt.Sprintf("templates/%s.tmpl", "mysql")
+func MongoDbMBuilderComposefile() error {
+	compose := setMongoDbSettings()
+	path := fmt.Sprintf("templates/%s.tmpl", "mongodb")
 
 	datafile, err := templatesContent.ReadFile(path)
 	if err != nil {
@@ -28,7 +28,7 @@ func MysqlBuilderComposefile() error {
 	return nil
 }
 
-func setMysqlSettings() *ComposeData {
+func setMongoDbSettings() *ComposeData {
 	data := new(ComposeData)
 
 	prompts := []struct {
@@ -37,12 +37,9 @@ func setMysqlSettings() *ComposeData {
 	}{
 		{"docker-compose Version (Default: 3.9)", &data.Version},
 		{"Image Version (Default: latest)", &data.ImageVersion},
-		{"DB Name (Default: DB)", &data.DbName},
-		{"DB User (Default: USER)", &data.DbUser},
-		{"DB Root Password (Default: ROOT)", &data.DbRootPass},
-		{"DB Password (Default: PASS)", &data.DbPass},
-		{"Restart (Default: no)", &data.Restart},
-		{"Ports (Default: 3306:3306)", &data.Ports},
+		{"MONGO_INITDB_ROOT_USERNAME (Default: USER)", &data.DbUser},
+		{"MONGO_INITDB_ROOT_PASSWORD (Default: PASS)", &data.DbPass},
+		{"Ports (Default: 27017:27017)", &data.Ports},
 		{"CPU (Default: 1)", &data.Cpu},
 		{"Memory (MB) (Default: 500)", &data.Memory},
 		{"Network Name (Default: network)", &data.NetworkName},
@@ -61,15 +58,11 @@ func setMysqlSettings() *ComposeData {
 
 		*p.Pointer = result
 	}
-
 	if len(data.Version) == 0 {
 		data.Version = "3.9"
 	}
 	if len(data.ImageVersion) == 0 {
 		data.ImageVersion = "latest"
-	}
-	if len(data.DbName) == 0 {
-		data.DbName = "DB"
 	}
 	if len(data.DbUser) == 0 {
 		data.DbUser = "USER"
@@ -77,11 +70,8 @@ func setMysqlSettings() *ComposeData {
 	if len(data.DbPass) == 0 {
 		data.DbPass = "PASS"
 	}
-	if len(data.Restart) == 0 {
-		data.Restart = "no"
-	}
 	if len(data.Ports) == 0 {
-		data.Ports = "3306:3306"
+		data.Ports = "27017:27017"
 	}
 	if len(data.Cpu) == 0 {
 		data.Cpu = "1"
@@ -91,9 +81,6 @@ func setMysqlSettings() *ComposeData {
 	}
 	if len(data.NetworkName) == 0 {
 		data.NetworkName = "network"
-	}
-	if len(data.DbRootPass) == 0 {
-		data.DbRootPass = "ROOT"
 	}
 
 	return data
