+++ internal/builders/compose/templates/mongodb.tmpl
