+++ internal/builders/compose/templates/mysql.tmpl
